Split the web api code in multiple files
